Fixed missed replacement of Config, moved more classes to output.dot.schemaspy
