+++ src/main/java/org/schemaspy/output/dot/schemaspy/DotNode.java
@@ -27,8 +27,6 @@ import org.schemaspy.model.Table;
 import org.schemaspy.model.TableColumn;
 import org.schemaspy.model.TableIndex;
 import org.schemaspy.output.dot.DotConfig;
-import org.schemaspy.util.Html;
-import org.schemaspy.view.StyleSheet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
