+++ src/main/java/org/schemaspy/output/dot/schemaspy/Html.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with SchemaSpy. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.schemaspy.util;
+package org.schemaspy.output.dot.schemaspy;
 
 public class Html {
