+++ src/main/java/org/schemaspy/output/dot/schemaspy/StyleSheet.java
@@ -20,12 +20,10 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.schemaspy.view;
+package org.schemaspy.output.dot.schemaspy;
 
 import org.schemaspy.Config;
 import org.schemaspy.model.InvalidConfigurationException;
-import org.schemaspy.util.ResourceFinder;
-import org.schemaspy.util.ResourceNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,7 +35,7 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.StringTokenizer;
 
-import static org.schemaspy.view.StyleSheetConst.*;
+import static org.schemaspy.output.dot.schemaspy.StyleSheetConst.*;
 
 /**
  * Represents our CSS style sheet (CSS) with accessors for important
